requirements.txt
anthropic>=0.42.0
openai>=1.0.0
streamlit==1.42.1
python-dotenv==1.0.1
pytest==7.4.4
